infra/dags/requirements.txt
apache-airflow-providers-amazon==6.0.0